refactor: src directory
