--- src/App.css
@@ -1,50 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.Navbar{
-  display: flex;
-  align-items: center;
-  padding-left: 20px;
-  height: 50px;
-  text-align: center;
-  color: yellowgreen;
-  background-color: #474747;
-  gap: 10px;  
-}
-
-.editor{
-  margin-top: 0px;
-}
-.run-btn{
-  position: relative;   
-  width: 80px;  
-  font-size: 22px;
-  font-weight: bold;
-  background-color: #afec3f;
-  border: none;
-  border-radius: 4px;
-  transition: 0.3s;
-  cursor: pointer;
-}
-.run-btn:active{
-  background-color: black;
-}
-
-.compile-box{
-  flex: 40%; 
-  gap:50px;   
-  display: flex;
-  flex-direction: row;
-  background-color: #242424;
-  border: 3px solid #1f65e6;
-  padding: 5px;  
-}
-
-@media(max-width:700px){
-  .compile-box{
-      flex-direction: column;
-      gap:30px;
-      padding:0px;
-  }
-}
