+++ src/components/NavbarComponents/FontSizeControl.js
